Rename lib pipes
