+++ project/libs/shared/pipes/.eslintrc.json
