+++ project/libs/shared/pipes/README.md
@@ -1,0 +1,7 @@
+# shared-pipes
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-pipes` to execute the unit tests via [Jest](https://jestjs.io).
