+++ project/libs/shared/pipes/project.json
@@ -1,21 +1,21 @@
 {
-  "name": "shared-shared-pipes",
+  "name": "shared-pipes",
   "$schema": "../../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/shared/shared-pipes/src",
+  "sourceRoot": "libs/shared/pipes/src",
   "projectType": "library",
   "targets": {
     "lint": {
       "executor": "@nrwl/linter:eslint",
       "outputs": ["{options.outputFile}"],
       "options": {
-        "lintFilePatterns": ["libs/shared/shared-pipes/**/*.ts"]
+        "lintFilePatterns": ["libs/shared/pipes/**/*.ts"]
       }
     },
     "test": {
       "executor": "@nrwl/jest:jest",
       "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
       "options": {
-        "jestConfig": "libs/shared/shared-pipes/jest.config.ts",
+        "jestConfig": "libs/shared/pipes/jest.config.ts",
         "passWithNoTests": true
       },
       "configurations": {
