+++ project/libs/shared/pipes/src/index.ts
