+++ project/libs/shared/pipes/src/lib/mongoid-validation.pipe.ts
