+++ project/libs/shared/pipes/tsconfig.json
