+++ project/libs/shared/pipes/tsconfig.lib.json
