+++ project/libs/shared/pipes/tsconfig.spec.json
