--- project/libs/shared/shared-pipes/README.md
@@ -1,7 +1,0 @@
-# shared-shared-pipes
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test shared-shared-pipes` to execute the unit tests via [Jest](https://jestjs.io).
