+++ project/tsconfig.base.json
@@ -24,7 +24,7 @@
       "@project/config/config-users": ["libs/config/config-users/src/index.ts"],
       "@project/models/models-blog": ["libs/models/models-blog/src/index.ts"],
       "@project/shared/app-types": ["libs/shared/app-types/src/index.ts"],
-      "@project/shared/shared-pipes": ["libs/shared/shared-pipes/src/index.ts"],
+      "@project/shared/pipes": ["libs/shared/pipes/src/index.ts"],
       "@project/util/util-core": ["libs/util/util-core/src/index.ts"],
       "@project/util/util-types": ["libs/util/util-types/src/index.ts"]
     }
